+++ fiddles/vue/fiddle-0018-InfiniteSelect/api/utils/response-throttle.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const DEFAULT_TIMEOUT = 1
+
+exports.throttle = (ms) => {
+    return new Promise(resolve => setTimeout(resolve, ms || DEFAULT_TIMEOUT))
+}